+++ srcs/app/backend/tests/refundable_escrow_test.ts
@@ -1,101 +1,9 @@
-import { PublicKey, LAMPORTS_PER_SOL, Keypair, SystemProgram } from "@solana/web3.js";
+import { LAMPORTS_PER_SOL, Keypair } from "@solana/web3.js";
 import * as anchor from "@coral-xyz/anchor";
-import { Program, BN, AnchorProvider } from "@coral-xyz/anchor";
-import { RefundableEscrow } from "../target/types/refundable_escrow";
-import { Buffer } from "buffer";
+import { BN } from "@coral-xyz/anchor";
 import { expect } from "chai";
-
-enum ResultStatus {
-	SUCCESS,
-	ERROR,
-}
-
-async function execCreateRefundableEscrow(
-	program: Program,
-	buyer: Keypair,
-	seller: Keypair,
-	escrow: RefundableEscrow,
-	transactionId: BN,
-	amountLamports: BN,
-	refundableSeconds: BN,
-	userDefinedData: String
-): Promise<ResultStatus> {
-	try {
-
-		await program.methods.createRefundableEscrow(
-			transactionId,
-			amountLamports,
-			refundableSeconds,
-			userDefinedData)
-			.accounts({
-				buyer: buyer.publicKey,
-				seller: seller.publicKey,
-				escrow: escrow,
-				systemProgram: SystemProgram.programId,
-			})
-			.signers([buyer])
-			.rpc();
-		return ResultStatus.SUCCESS;
-	} catch {
-		return ResultStatus.ERROR;
-	}
-}
-
-async function execSettleLamports(
-	program: Program,
-	requestor: Keypair,
-	escrow: RefundableEscrow): Promise<ResultStatus> {
-	try {
-		await program.methods.settleLamports()
-			.accounts({
-				requestor: requestor.publicKey,
-				escrow: escrow
-			})
-			.signers([requestor])
-			.rpc();
-		return ResultStatus.SUCCESS;
-	} catch {
-		return ResultStatus.ERROR;
-	}
-}
-
-async function getEscrowPDA(
-	buyer: Keypair,
-	seller: Keypair,
-	transactionId: BN,
-	programId: PublicKey) {
-	// convert to bytes
-	const transactionIdBuffer = Buffer.alloc(8);
-	transactionIdBuffer.writeBigInt64LE(BigInt(transactionId.toNumber()), 0);
-
-	const seeds = [
-		buyer.publicKey.toBuffer(),
-		seller.publicKey.toBuffer(),
-		transactionIdBuffer,
-	]
-
-	const [escrowPDA, _] = await PublicKey.findProgramAddress(
-		seeds,
-		programId
-	);
-	return escrowPDA;
-}
-
-async function airdropLamports(
-	provider: AnchorProvider,
-	keypair: Keypair,
-	amountLamports: number) {
-	const connection = provider.connection;
-	const signature = await connection.requestAirdrop(keypair.publicKey, amountLamports);
-	await connection.confirmTransaction(signature);
-}
-
-function sleep(seconds: number) {
-	const milliseconds = seconds * 1000;
-	const start = Date.now();
-	while (Date.now() - start < milliseconds) {
-	}
-}
+import { execCreateRefundableEscrow, execSettleLamports, ResultStatus } from "./escrow";
+import { sleep, getEscrowPDA, airdropLamports } from "./utils";
 
 describe("Refundable Escrow Program", () => {
 	const provider = anchor.AnchorProvider.env();
